ether: wifi: Switch to duty cycling QPower
This addresses the wifi instability.
The problem is that the full QPower mode
(value 2) is too aggressive and causes
the driver to be in a bad state even when
it is needed.
As solution, I've tested the value 1 which
represents the legacy power management
mechanics and the value 5 which represents
duty cycling QPower.
Both fixed the stability issue.
I am choosing duty cycling QPower since
it has a smart logic which toggles
QPower based on an inactivity timer
and has a better power impact than
the legacy power management mechanics.
That matches the value Google uses
for Bullhead.

For surya, value of 4 seems to be the preferred one.

Change-Id: Ic78646215ef51344a46d10dd41b458bc99fbc06e