surya: added refresh rate settings
Added refresh rate setting to select refresh rate, including a quick
settings tile.

Dynamic refresh rate is not yet implemented on Android 10, but the
device defaults to 120Hz refresh rate which is a battery drain. Stock
ROM allows for selection of lower refresh rates, this commit adds the
same capability.

Change-Id: Ia7e736211f9d478476bc25dbb6f12767aab1ed12